safari-600.7-branch/LayoutTests/platform/win/editing/style/remove-underline-from-stylesheet-expected.txt

This tests removing underline from stylesheet (.editing has underline). Because text-decoration that comes from a style rule cannot be pushed down, the underline should stay. (see bug 27809).
| "xxxxxx "
| "<#selection-anchor>xxxxxx <#selection-focus>"
| "xxxxxx"
| <span>
|   id="test"
